+++ CHANGELOG.md
@@ -15,6 +15,9 @@ See [docs/RELEASING.md](docs/RELEASING.md).
 
 ### Enhancements
 
+- **Power Templates:** renamed from PDU Templates; page has separate **PDU templates** and **UPS templates** cards.
+- **UPS inventory templates:** create from UPS detail (**Create UPS template**), manage under Power Templates, select on **New UPS** to prefill model/size/SNMP/OID map.
+
 ### Bug fixes
 
 ---